refactor(update): dedupe shared helpers into @loadout/types
Dead-code / duplication audit of the PR surface. Every export is
consumed by production code or an exercised test seam; the real
duplication was isTrustedGithubHost, parseSha256Sums and makeIdleAbort
copied between the loader and overlay updaters (with duplicated
tests), plus host.ts re-declaring the UpdateStatus/UpdateCheckResult
shapes updater.ts defines. The store-bridge convention ("copy until a
third consumer appears") hit its threshold, so the pure helpers, the
watchdog, and the RPC-boundary status types now live in
@loadout/types/update-shared with a single test file; both updaters
import them and host.ts re-exports the types.
