+++ apps/loadout/src/loader/self-update.ts
@@ -29,40 +29,19 @@ import { chmod, copyFile, mkdir, rename, rm, stat, statfs } from "node:fs/promis
 import { readlinkSync } from "node:fs";
 import { basename, dirname, join } from "node:path";
 import { runFull, commandExists, spawn } from "@loadout/exec";
-import { RELEASE_TAG_RE, parseVersion, compareVersions } from "@loadout/types";
+import {
+  RELEASE_TAG_RE,
+  parseVersion,
+  compareVersions,
+  isTrustedGithubHost,
+  parseSha256Sums,
+  makeIdleAbort,
+} from "@loadout/types";
 import { LOADER_VERSION } from "../version";
 import { getTargetUser } from "./target-user";
 import { log } from "./logger";
 
 const REPO = "srsholmes/loadout";
-
-/** Hosts a release-asset redirect chain may pass through. Checked on
- *  EVERY hop (same rationale as plugins/store-bridge/lib/github-release.ts:
- *  an attacker-controlled hop must fail before its body is fetched). */
-const TRUSTED_GITHUB_HOSTS = [
-  "github.com",
-  "objects.githubusercontent.com",
-  "release-assets.githubusercontent.com",
-];
-
-export function isTrustedGithubHost(host: string): boolean {
-  const lower = host.toLowerCase();
-  return TRUSTED_GITHUB_HOSTS.some((t) => lower === t || lower.endsWith(`.${t}`));
-}
-
-/** Parse a `sha256sum`-format manifest ("<hex>  <filename>" lines)
- *  into filename → lowercase hex. Tolerates the `*` binary-mode
- *  marker and blank lines. */
-export function parseSha256Sums(text: string): Map<string, string> {
-  const out = new Map<string, string>();
-  for (const line of text.split("\n")) {
-    const m = /^([0-9a-fA-F]{64})\s+\*?(.+)$/.exec(line.trim());
-    const hash = m?.[1];
-    const name = m?.[2];
-    if (hash && name) out.set(name.trim(), hash.toLowerCase());
-  }
-  return out;
-}
 
 export type SelfUpdatePhase = "idle" | "downloading" | "verifying" | "applying" | "done" | "error";
 
@@ -175,40 +154,11 @@ function assetHopHeaders(host: string): Record<string, string> {
 }
 
 /** How long a download may go without a single received byte before
- *  it's aborted — a half-open TCP connection would otherwise hang
- *  `fetch` forever, pinning the status at `downloading` where every
- *  new POST is refused with 409 until the service restarts. */
+ *  the makeIdleAbort watchdog (from @loadout/types) kills it — a
+ *  half-open TCP connection would otherwise hang `fetch` forever,
+ *  pinning the status at `downloading` where every new POST is
+ *  refused with 409 until the service restarts. */
 const DOWNLOAD_IDLE_TIMEOUT_MS = 60_000;
-
-/** AbortController that fires after `ms` of silence; reset() on every
- *  received chunk. Timers are unref'd so a pending watchdog never
- *  keeps the process alive on its own. */
-function makeIdleAbort(ms: number): {
-  signal: AbortSignal;
-  reset: () => void;
-  clear: () => void;
-} {
-  const controller = new AbortController();
-  const arm = () => {
-    const t = setTimeout(
-      () => controller.abort(new Error(`download stalled (no data for ${ms / 1000}s)`)),
-      ms,
-    );
-    (t as unknown as { unref?: () => void }).unref?.();
-    return t;
-  };
-  let timer = arm();
-  return {
-    signal: controller.signal,
-    reset() {
-      clearTimeout(timer);
-      timer = arm();
-    },
-    clear() {
-      clearTimeout(timer);
-    },
-  };
-}
 
 export async function downloadToFile(
   url: string,
